debian:stretch-slim

+++ Dockerfile
@@ -1,22 +1,34 @@
-#FROM nginx:stable
+FROM debian:stretch-slim
 
-FROM debian:jessie
+MAINTAINER Positron <positron@jarm.com>
 
-MAINTAINER NGINX Docker Maintainers "docker-maint@nginx.com"
+ENV NGINX_VERSION 1.12.0-1~stretch
+ENV NJS_VERSION   1.12.0.0.1.10-1~stretch
 
-ENV NGINX_VERSION 1.12.0-1~jessie
-
-RUN apt-key adv --keyserver hkp://pgp.mit.edu:80 --recv-keys 573BFD6B3D8FBC641079A6ABABF5BD827BD9BF62 \
-	&& echo "deb http://nginx.org/packages/debian/ jessie nginx" >> /etc/apt/sources.list \
+RUN apt-get update \
+	&& apt-get install --no-install-recommends --no-install-suggests -y gnupg1 \
+	&& \
+	NGINX_GPGKEY=573BFD6B3D8FBC641079A6ABABF5BD827BD9BF62; \
+	found=''; \
+	for server in \
+		ha.pool.sks-keyservers.net \
+		hkp://keyserver.ubuntu.com:80 \
+		hkp://p80.pool.sks-keyservers.net:80 \
+		pgp.mit.edu \
+	; do \
+		echo "Fetching GPG key $NGINX_GPGKEY from $server"; \
+		apt-key adv --keyserver "$server" --keyserver-options timeout=10 --recv-keys "$NGINX_GPGKEY" && found=yes && break; \
+	done; \
+	test -z "$found" && echo >&2 "error: failed to fetch GPG key $NGINX_GPGKEY" && exit 1; \
+	apt-get remove --purge -y gnupg1 && apt-get -y --purge autoremove && rm -rf /var/lib/apt/lists/* \
+	&& echo "deb http://nginx.org/packages/debian/ stretch nginx" >> /etc/apt/sources.list \
 	&& apt-get update \
 	&& apt-get install --no-install-recommends --no-install-suggests -y \
-						ca-certificates \
 						nginx=${NGINX_VERSION} \
-						nginx-module-xslt \
-						nginx-module-geoip \
-						nginx-module-image-filter \
-						nginx-module-perl \
-						nginx-module-njs \
+						nginx-module-xslt=${NGINX_VERSION} \
+						nginx-module-geoip=${NGINX_VERSION} \
+						nginx-module-image-filter=${NGINX_VERSION} \
+						nginx-module-njs=${NJS_VERSION} \
 						gettext-base \
 	&& rm -rf /var/lib/apt/lists/*
 
@@ -24,9 +36,11 @@ RUN apt-key adv --keyserver hkp://pgp.mit.edu:80 --recv-keys 573BFD6B3D8FBC64107
 RUN ln -sf /dev/stdout /var/log/nginx/access.log \
 	&& ln -sf /dev/stderr /var/log/nginx/error.log
 
-#EXPOSE 80 443
+#EXPOSE 80
 
-CMD ["nginx", "-g", "daemon off;"]
+#STOPSIGNAL SIGQUIT
+
+#CMD ["nginx", "-g", "daemon off;"]
 
 
 RUN mkdir -p /var/www \
@@ -39,3 +53,7 @@ COPY nginx.conf /etc/nginx/
 VOLUME ["/var/www"]
 
 EXPOSE 80 81 82 443
+
+STOPSIGNAL SIGQUIT
+
+CMD ["nginx", "-g", "daemon off;"]
